chatbot/requirements.txt
streamlit>=1.35.0
transformers>=4.40.0
torch>=2.0.0
huggingface_hub>=0.20.0
requests>=2.28.0
python-dotenv>=1.0.0
ibm-watsonx-ai>=1.0.0
ibm-watson>=8.0.0
groq>=0.4.0
pandas>=1.5.0,<2.2.0
numpy>=1.24.0
plotly>=5.20.0
streamlit-chat>=0.1.0
